+++ src/CLI/Command.hs
@@ -176,6 +176,12 @@ executeCommand (CLIConfig globalOpts _) cmd manager = do
         ExportMatrices _ -> do
             (database, _solver) <- requireDatabase manager (dbName globalOpts)
             executeDbCommand outputFormat globalOpts database cmd
+        QualityReport _ -> do
+            reportError "quality-report is served over HTTP; Main.hs routes it to the client"
+            exitFailure
+        ComputedQualityReport _ -> do
+            reportError "computed-quality-report is served over HTTP; Main.hs routes it to the client"
+            exitFailure
         Stop -> do
             reportError "Stop command should be handled in Main.hs"
             exitFailure
@@ -214,6 +220,8 @@ executeDbCommand fmt _globalOpts database = \case
     CompartmentMappings -> pure ()
     Units -> pure ()
     FlowMapping _ -> pure ()
+    QualityReport _ -> pure ()
+    ComputedQualityReport _ -> pure ()
     Stop -> pure ()
     Repl -> pure ()
     Dump _ -> pure ()